+++ pronunciation_master/src/process_mongodb_ipa.py
@@ -8,15 +8,20 @@ from pymongo import MongoClient
 
 
 
-def process_db(database):
+def process_db(database, word=None):
     client = MongoClient()
     db = getattr(client, database)
-    db.wiktionary_ipa.drop()
-    for document in db.wiktionary_raw_subdivided.find({}):
+    if not word:
+        db.wiktionary_ipa.drop()
+    query = word and {'word': word} or {}
+    for document in db.wiktionary_raw_subdivided.find(query):
         if not 'section' in document:
             import pdb; pdb.set_trace()
         if document['section'] == 'Pronunciation':
             ipa = process_ipa(document['content'])
+            if word:
+                print ipa
+                print document
             db.wiktionary_ipa.insert_one({
                 'language': document['language'],
                 'word': document['word'],
@@ -26,12 +31,13 @@ def process_db(database):
 
 
 def process_ipa(content):
-    pronunciations = re.findall('IPA\|/(.*)/\|', content)
+    pronunciations = re.findall('{{.*IPA\|/?([^/|\n]*)/?\|?.*}}', content)
     return pronunciations
 
 
 if __name__ == '__main__':
     parser = argparse.ArgumentParser()
     parser.add_argument('--database', required=True)
+    parser.add_argument('--word', required=False)
     args = parser.parse_args()
-    process_db(args.database)
+    process_db(args.database, args.word)
